경로 조회 api 성능 개선
