+++ src/common/geo/converter.d.ts
@@ -1,0 +1,3 @@
+import type { Converter } from "proj4";
+declare const converter: Converter;
+export = converter;
