+++ src/common/geo/converter.js
@@ -1,8 +1,9 @@
-import proj from "proj4";
+const proj = require("proj4");
 
 const __WGS84 = "EPSG:4326";
 const __UTM_K = "EPSG:5179";
 const __PROJ = "+proj=tmerc +lat_0=38 +lon_0=127.5 +k=0.9996 +x_0=1000000 +y_0=2000000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs";
 
 proj.defs(__UTM_K, __PROJ);
-export default proj(__WGS84, __UTM_K);
+const converter = proj(__WGS84, __UTM_K);
+module.exports = converter;
