--- src/config/workerpool/index.ts
@@ -1,2 +1,0 @@
-export * from "./workerpool.module";
-export * from "./workerpool.service";
