--- src/config/workerpool/token.ts
@@ -1,2 +1,0 @@
-
-export const WORKER_POOL_OPTIONS = Symbol("WORKER_POOL_OPTIONS");
