--- src/config/workerpool/workerpool.options.factory.ts
@@ -1,12 +1,0 @@
-import { ConfigService } from "@nestjs/config";
-import { WorkerPoolOptions } from "workerpool";
-
-export function workerPoolOptionsFactory(
-    config: ConfigService
-): WorkerPoolOptions {
-    return {
-        workerType: "thread",
-        maxWorkers: config.get<number>("POOL_MAX_WORKERS") ?? 5,
-        maxQueueSize: config.get<number>("POOL_MAX_QUEUE_SIZE") ?? 30,
-    };
-}
