+++ src/courses/courses.module.ts
@@ -10,13 +10,8 @@ import { CoursesController } from "./controller";
 import { RunningRecord } from "../modules/running";
 import { RunningDashboardsService } from "../running/service";
 import { RunningModule } from "../running/running.module";
-import { WorkerPoolModule, WorkerPoolService } from "../config/workerpool";
-import { resolve } from "node:path";
 
-const __EXTERNAL_PROVIDERS = [
-    WorkerPoolService,
-    RunningDashboardsService
-];
+const __EXTERNAL_PROVIDERS = [RunningDashboardsService];
 
 @Module({
   imports: [
@@ -26,7 +21,6 @@ const __EXTERNAL_PROVIDERS = [
           CourseBookmark,
           RunningRecord,
       ]),
-      WorkerPoolModule.register(resolve(__dirname, "/worker/inspect.path.worker.js")),
       forwardRef(() => RunningModule),
   ],
   providers: [
